생산계획 수량 소수점 지원 수정
- batch_create에서 quantity .to_i → .to_f 변경
- 기정떡 0.5통 단위 저장 가능하도록 수정

+++ app/controllers/production/plans_controller.rb
@@ -22,7 +22,7 @@ class Production::PlansController < ApplicationController
 
     ActiveRecord::Base.transaction do
       plans_data.each do |recipe_id, plan_data|
-        quantity = plan_data[:quantity].to_i
+        quantity = plan_data[:quantity].to_f  # 기정떡은 0.5 단위 지원
         unit_type = plan_data[:unit_type].presence || '개'
         allocations = plan_data[:allocations] || {}
